refactor: move functions into a single file
