+++ changelog.md
@@ -2,5 +2,6 @@
 
 ## Latest Changes
 
+- refactor: move functions into a single file
 - docs: switch to google type docstrings
 - docs: switch from sphinx to mkdocs
