+++ docs/user-guide/tools/signal.md
@@ -1,4 +1,4 @@
-::: pysmo.tools.noise
+::: pysmo.tools.signal
     options:
       show_root_toc_entry: false
       show_root_heading: true
