--- pysmo/functions/README.md
@@ -1,3 +1,0 @@
-# Functions folder
-
-This is where pysmo functions live. A `__init__.py`  file is used here to manage them. Thus new functions should be added so that they can be imported via `pysmo.functions`.
