--- pysmo/functions/__init__.py
@@ -1,23 +1,0 @@
-"""
-Pysmo provides functions that perform common operations on the types of data that match
-pysmo's types.
-"""
-
-from .resample import resample
-from .filter_gauss import envelope, gauss
-from .plotseis import plotseis
-from .azdist import azimuth, backazimuth, distance
-from .normalize import normalize
-from .detrend import detrend
-
-__all__ = [
-    'detrend',
-    'resample',
-    'envelope',
-    'gauss',
-    'plotseis',
-    'azimuth',
-    'backazimuth',
-    'distance',
-    'normalize'
-]
